test(libeval): split redaction-pipeline by behaviour family
Split redaction-pipeline.test.js (506 LOC) into
redaction-pipeline-producer.test.js (AgentRunner producer-side: sentinel
sweep, patterns, opt-out, toText fidelity) and
redaction-pipeline-orchestrator.test.js (Supervisor/Facilitator summary
redaction). Shared sink/sentinel setup lifted into
redaction-pipeline-helpers.js. Assertions unchanged; 7 tests preserved.

Spec 0640 § C (large files), design Decision 5.
